move agentstream_de to base agentstream branch
git-svn-id: http://cubeless.dev.sabre.com:18080/svn/cubeless/communities/agentstream@19803 e05bbbe3-11a3-40ec-a18c-f20a9e1084ec

+++ app/models/topic.rb
@@ -8,7 +8,7 @@ class Topic
       # process the indexing in dalay job
       ChatTopicIndex.delay.update_indices(self) unless self.chat.on_air?
     else
-      chat_topic_index.delay.destroy  
+      chat_topic_index.delay.destroy if chat_topic_index
     end
   end
